# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse_null_": {
				"properties": {
					"data": {
						"items": {
							"type": "number",
							"enum": [
								null
							],
							"nullable": true
						},
						"type": "array",
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"enum": [
					"ADMIN",
					"SUPER_ADMIN",
					"VENDOR",
					"CUSTOMER"
				],
				"type": "string"
			},
			"CustomerProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"shippingAddress": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"preferences": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VendorProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessRegistrationNumber": {
						"type": "string"
					},
					"businessAddress": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"totalReviews": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"businessName",
					"isVerified",
					"rating",
					"totalReviews"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfileResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"isVendorVerified": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"phoneNumber": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessAddress": {
						"type": "string"
					},
					"customer": {
						"$ref": "#/components/schemas/CustomerProfileDto"
					},
					"vendor": {
						"$ref": "#/components/schemas/VendorProfileDto"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"role",
					"isEmailVerified",
					"isVendorVerified",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserProfileResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/UserProfileResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/UserProfileResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AuthResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserProfileResponse"
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AuthResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/AuthResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/AuthResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginSchema": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TokenResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TokenResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/TokenResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenSchema": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordSchema": {
				"properties": {
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordSchema": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordSchema": {
				"properties": {
					"token": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AdminUpdateUserStatusSchema": {
				"properties": {
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminVerifyVendorSchema": {
				"properties": {
					"isVerified": {
						"type": "boolean"
					}
				},
				"required": [
					"isVerified"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminPermission": {
				"enum": [
					"PRODUCT",
					"LETTERS",
					"CONTACTS",
					"LOGS"
				],
				"type": "string"
			},
			"AdminLoginResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/AdminPermission"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"role",
					"permissions",
					"isActive",
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AdminLoginResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/AdminLoginResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/AdminLoginResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminAuthSchema": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_RefreshTokenResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/RefreshTokenResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/RefreshTokenResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminRefreshTokenSchema": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminProfileResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/AdminPermission"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"role",
					"permissions",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AdminProfileResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/AdminProfileResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/AdminProfileResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminChangePasswordSchema": {
				"properties": {
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminForgotPasswordSchema": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminResetPasswordSchema": {
				"properties": {
					"token": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Otherly API",
		"version": "1.0.0",
		"description": "Backend REST API for Otherly reverse-commerce marketplace"
	},
	"paths": {
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginSchema"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TokenResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenSchema"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "Me",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordSchema"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordSchema"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordSchema"
							}
						}
					}
				}
			}
		},
...
		"/admin/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pagination": {
											"properties": {
												"lastPage": {
													"type": "number",
													"format": "double"
												},
												"prevPage": {
													"type": "number",
													"format": "double"
												},
												"nextPage": {
													"type": "number",
													"format": "double"
												},
												"currentPage": {
													"type": "number",
													"format": "double"
												},
												"count": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"lastPage",
												"prevPage",
												"nextPage",
												"currentPage",
												"count"
											],
											"type": "object"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/UserProfileResponse"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"pagination",
										"data",
										"success",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Role"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isVendorVerified",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/admin/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/users/{id}/status": {
			"patch": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminUpdateUserStatusSchema"
							}
						}
					}
				}
			}
		},
		"/admin/users/{id}/verify-vendor": {
			"patch": {
				"operationId": "VerifyVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminVerifyVendorSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AdminLoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminAuthSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_RefreshTokenResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminRefreshTokenSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/auth/me": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AdminProfileResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminChangePasswordSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminForgotPasswordSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminResetPasswordSchema"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:4000/api/v1",
			"description": "Development server"
		}
	]
}